peremoga 70 char len
