+++ WM/Symbol
@@ -1,73 +1,40 @@
-"........................................"
-"........................................"
-"...................#*..................."
-"..................*##*.................."
-".....##...........####..........:##....."
-".....####.........####........:####....."
-".....#####*.......####.......######....."
-".....##*.###......###*......###.###....."
-".....##*..##*.....*##*.....*##..###....."
-".....##*..*##.....*##*.....##*..###....."
-".....##*..:##:....*##*....:##:..###....."
-".....##*...##*....*##*....*##...###....."
-".....##*...###....####....###...###....."
-".....##*...*##....####....##*...###....."
-".....##*..####...#####*..:####..###....."
-".....##*.###....*##::##:....###.###....."
-".....##*:##:...:##*..*##:...:##.###....."
-".....##*.*###.:##:....*##::###*.###....."
-".....##*...:::##:*#::#:*##.::...###....."
-".....##*.....###.:####:.##*.....###....."
-".....###*********:*##::*********###....."
-".....############:*##:*############....."
-"..............###.*##:.###.............."
-"...............#########*..............."
-".................######................."
-"...................::..................."
-
-"__       __  __"
-"/  |  _  /  |/  |"
-"$$ | / \ $$ |$$/  _______   _______    ______    ______"
-"$$ |/$  \$$ |/  |/       \ /       \  /      \  /      \"
-"$$ /$$$  $$ |$$ |$$$$$$$  |$$$$$$$  |/$$$$$$  |/$$$$$$  |"
-"$$ $$/$$ $$ |$$ |$$ |  $$ |$$ |  $$ |$$    $$ |$$ |  $$/"
-"$$$$/  $$$$ |$$ |$$ |  $$ |$$ |  $$ |$$$$$$$$/ $$ |"
-"$$$/    $$$ |$$ |$$ |  $$ |$$ |  $$ |$$       |$$ |"
-"$$/      $$/ $$/ $$/   $$/ $$/   $$/  $$$$$$$/ $$/"
-
-" ______   __        __            __"
-"/      \ /  |      /  |          /  |"
-"/$$$$$$  |$$ |____  $$/   _______ $$ |   __   ______   _______"
-"$$ |  $$/ $$      \ /  | /       |$$ |  /  | /      \ /       \"
-"$$ |      $$$$$$$  |$$ |/$$$$$$$/ $$ |_/$$/ /$$$$$$  |$$$$$$$  |"
-"$$ |   __ $$ |  $$ |$$ |$$ |      $$   $$<  $$    $$ |$$ |  $$ |"
-"$$ \__/  |$$ |  $$ |$$ |$$ \_____ $$$$$$  \ $$$$$$$$/ $$ |  $$ |"
-"$$    $$/ $$ |  $$ |$$ |$$       |$$ | $$  |$$       |$$ |  $$ |"
-"$$$$$$/  $$/   $$/ $$/  $$$$$$$/ $$/   $$/  $$$$$$$/ $$/   $$/"
-
-" _______   __"
-" /       \ /  |"
-"$$$$$$$  |$$/  _______   _______    ______    ______"
-"$$ |  $$ |/  |/       \ /       \  /      \  /      \"
-"$$ |  $$ |$$ |$$$$$$$  |$$$$$$$  |/$$$$$$  |/$$$$$$  |"
-"$$ |  $$ |$$ |$$ |  $$ |$$ |  $$ |$$    $$ |$$ |  $$/"
-"$$ |__$$ |$$ |$$ |  $$ |$$ |  $$ |$$$$$$$$/ $$ |"
-"$$    $$/ $$ |$$ |  $$ |$$ |  $$ |$$       |$$ |"
-"$$$$$$$/  $$/ $$/   $$/ $$/   $$/  $$$$$$$/ $$/"
-
-"#######"
-"##   ##"
-"##   ##  #####  ######   #####  ##   ##  #####   #####   ####"
-"##   ## ##   ## ##   ## ##   ## ### ### ##   ##      ##     ##"
-"##   ## ######  ######  ######  ## # ## ##   ##  #####   #####"
-"##   ## ##      ##      ##      ##   ## ##   ## ##      ##  ##"
-"##   ##  #####  ##       #####  ##   ##  #####   ######  ######"
+"                                        "
+"                                        "
+"                   #*                   "
+"                  *##*                  "
+"     ##           ####           ##     "
+"     ####         ####         ####     "
+"     #####*       ####       ######     "
+"     ##* ###      ###*      ### ###     "
+"     ##*  ##*     *##*     *##  ###     "
+"     ##*  *##     *##*     ##*  ###     "
+"     ##*   ##     *##*     ##   ###     "
+"     ##*   ##*    *##*    *##   ###     "
+"     ##*   ###    ####    ###   ###     "
+"     ##*   *##    ####    ##*   ###     "
+"     ##*  ####   #####*   ####  ###     "
+"     ##* ###    *##  ##     ### ###     "
+"     ##* ##     ##*  *##     ## ###     "
+"     ##* *###  ##     *##  ###* ###     "
+"     ##*      ## *#  # *##      ###     "
+"     ##*     ###  ####  ##*     ###     "
+"     ###********* *##  *********###     "
+"     ############ *## *############     "
+"              ### *##  ###              "
+"               #########*               "
+"                 ######                 "
+"                                        "
 
 
-"########  #######  ######   #######  ##     ##   #####   #######     ####"
-"##    ##  ##       ##   ##  ##       ##     ##  ##   ##  ##         ## ##"
-"##    ##  ##       ##   ##  ##       #### ####  ##   ##  ##        ##  ##"
-"##    ##  ######   ######   ######   ##  #  ##  ##   ##  ##       ##   ##"
-"##    ##  ##       ##       ##       ##     ##  ##   ##  ##       #######"
-"##    ##  ##       ##       ##       ##     ##  ##   ##  ##       ##   ##"
-"##    ##  #######  ##       #######  ##     ##   #####   ##       ##   ##"
+
+"#######  ######  ######   ######  ##     ##   #####   ######   ####"
+"##   ##  ##      ##   ##  ##      ###   ###  ##   ##  ##      ## ##"
+"##   ##  ##      ##   ##  ##      ## # # ##  ##   ##  ##     ##  ##"
+"##   ##  #####   ######   #####   ##  V  ##  ##   ##  ##    ##   ##"
+"##   ##  ##      ##       ##      ##     ##  ##   ##  ##    #######"
+"##   ##  ##      ##       ##      ##     ##  ##   ##  ##    ##   ##"
+"##   ##  ######  ##       ######  ##     ##   #####   ##    ##   ##"
+
+
+
+
